# Replace scale algorithm in Image::scale