There's a long discussion of where this came from and why in issue #997 .

I've suggested several alternatives there, and propose this as my preference since it is a cleaner and shorter algorithm, fairly standard and well know and reliable as algorithms go (obviously flaws in my coding are possible), and uses the same for up and down.

I think this works as shown, but it could use a very hard look and people's opinions.

PS. There is one slightly unrelated change, near the top -- in case things are scaling up or down but very close to 100% (as more dymanic programs like mobile apps and alternative views might), this gives a bit of a boost.  A shame to run through the whole scale for 98% for example, just let the browser do any fine tuning.