@knnniggett 2015-08-22
I'm not sure I can agree that the slight benefit this PR brings to the image quality is worth using an algorithm that, according to #997 , is twice as slow as what we currently have.

Perhaps if the difference was more dramatic or the algorithm was not significantly slower.

Looking at the sample images, I still can't say I can clearly see a difference.  

Further discussion on this is welcome, especially from the other team members.