@Linwood-F 2015-08-22
Let me add a couple of points.

First, the motivation for the change is that the old algorithm is incorrect.  It does not accurately track array bounds as it iterates, which is what has caused the errors in allocation in the past, and the skewing of images at certain magnifications.  So the motivation was to make it actually correct, and not allow run-amok array bounds.

Secondly, the slowdown is most noticeable in large enlargements. Both algorithms scale exponentially, so relatively small underlying differences are magnified as you go from 100% to 200% to 400%, etc.

It is not clear whether this is slower or not (or by how much) in scaling down.  My GUESS Is most scaling in ZoneMinder is down, not up.  Perhaps some of the longer term users can comment there.  My speed testing was done as an average over a complete run from something like 10% to 400% by 1% increments.  With exponential increase, this means the average is weighted almost entirely to the very high end.  It would be worth trying to arrange a more isolated test (not involving ZMS and web display) and testing it in smaller scales, say 25% to 150% and see individually how it does.  I will try to do that.

Finally, I think it is work explaining the algorithm difference a bit. In the prior zoneminder scale up algorithm, it is written as some kind of a push loop, i.e. it loops through the old image by bit, and decides where to push data from that bit into the new image's one or more bits. Since the ratio is almost never integral, it has to round to deal with fractional bits, and it is in this rounding that it is going wrong.  The proposed algorithm is a normal resampling algorithm, which pulls data for each bit in the new image, making it more understandable (for whatever that is worth) and also more simple to ensure it honors array boundaries.  And it provides the same mechanism for up or down (again, more about human understanding).

My GUESS by the way, is if you eventually want to move toward library routines you are going to find this kind of choice frequently -- hand written, purpose-written code may be faster than a packaged, proven algorithm.  Now THIS is not a packaged solution (though it is an implementation of a proven algorithm, but it could contain coding errors).  But this particular discussion is going to resurface when you do incorporate packaged algorithms, as I doubt seriously they will not bring with them more overhead.

If I can isolate and test I will post some numbers.