@Linwood-F 2015-08-22
OK, I have actual times.  These are run on a HyperV VM so it is hard to say exactly how this compares to any given system, but for reference the VM will run 6 cameras at 2048 x 1536 and be about 40% busy.  These times were taken without zoneminder running.

The test was to take a 1024 x 768 JPG image and do 100 scales at various size, timing the loop, multiply by 10 to get milliseconds per scale (including one image copy overhead, which falls out at the 100% line). 

Scale (%)|New Time (ms)|Old Time (ms)|Increase
---------|-------------|-------------|--------
25|2.71|1.78|52%
50|6.26|1.54|306%
75|13.60|2.77|391%
100|0.39|0.38|<< no scale, overhead only
125|35.85|4.81|645%
150|52.12|5.59|832%
175|72.40|6.61|996%
200|94.64|7.36|1187%
225|117.20|8.61|1261%
250|143.35|10.03|1330%

As you can see, the overhead of ZMS and the web was hiding the actual increase, it is much larger than 3x.  On the other hand, the actual scale time is in small milliseconds when scaling down.  13ms (the 75%) supports 77 fps, but it is indeed almost 4x as much elapsed time (basically pure CPU) as the old program at just shy of 3ms.

The actual scale times go down with file size also.  My GUESS is that a 1024x768 image is never scaled up in real practice, but it is small VGA (etc.) images that may be scaled up.  I did the test again with a 640x480 image, and it scales like this: 25% = 1.7ms, 75% = 5.2ms, 150% = 20ms, 250%=56ms.   Comare the 56ms to the 143ms, and you can see it get better, but it still is long in comparison to the old algorithm.

I tend to be one that thinks you don't swap fast, incorrect code for slower correct code.  However, I identified over in another note a way I think you can prevent the bounds overrun in the old code (it does not make it right, it just makes it safer).  If you guys prefer that, it's easy to swap out.

We could also probably turn this into a nearest neighbor pull sample, though I doubt it makes a huge difference (they are the same order, just a bit less math for each bit). 

And someone better at optimizing C++ might want to look at the code for any obvious changes, like swapping some data types for faster math.

I still feel you are going to come back to this discussion again when someone actually integrates an image library to do these routine functions.  But I get that seeing 1330% is an eye catching number.