@knnniggett 2015-10-01
@Linwood-F Haven't seen you on our IRC channel lately, so I'll post here. I'm not ignoring this issue, but I simply have had other priorities.  I agree there is likely something not quite right with the way zonemidner currently handles this.  Sometimes working towards something, only to find that is not the way we want to do it, is still knowledge gained. Your contributions are certainly appreciated.