@knnniggett 2016-02-04
@Linwood-F I agree with the underlying thought that there is room for improvement in this area. However, I'm going to close this pr out for housekeeping sake, since increases in cpu requirements trump the benefits this pr would bring. Feel free create a new pr or new issue if you've got a new method.