@Linwood-F 2016-02-04
Do you want to take the range protection fix instead that was mentioned in the first comment in issue #997