@knnniggett 2016-02-04
I guess I misremembered. Thought resizing was necessary for the range protection which was causing the cpu spike. Yeah, go ahead and create a new pr if you can with the range protection fix.