@Linwood-F 2016-02-04
No, the scaling was done to use a standard library routine instead of a home grown (and somewhat flawed) one.  And it did improve the image slightly, but it was slower.

I'll do this, it will take me a bit to get back to programming as have some commitments coming up.  The code is actually in that issue but setting up and testing it will take me a bit of time, but will.  I need to bring a test system up to date anyway and reconcile all the last few bazillion commits.